Re-organized Begin/End, and refactored DicomFile/Directory Begin/EndOpen. Issue #65
