+++ DICOM/DICOM.csproj
@@ -64,6 +64,8 @@
     <Compile Include="DatabaseQueryTransformRule.cs" />
     <Compile Include="DicomDatasetWalkerExtensions.cs" />
     <Compile Include="AsyncFactory.cs" />
+    <Compile Include="DicomFile.Async.cs" />
+    <Compile Include="DicomFileExtensions.Async.cs" />
     <Compile Include="Imaging\DicomOverlayDataFactory.cs" />
     <Compile Include="Imaging\Mathematics\RectF.cs" />
     <Compile Include="IO\DesktopPath.cs" />
@@ -103,6 +105,7 @@
     <Compile Include="Log\Logger.cs" />
     <Compile Include="Log\LogLevel.cs" />
     <Compile Include="Log\LogManager.cs" />
+    <Compile Include="Media\DicomDirectory.Async.cs" />
     <Compile Include="Media\DicomDirectoryReaderObserver.cs" />
     <Compile Include="Media\DicomFileScanner.cs" />
     <Compile Include="DicomMatchRules.cs" />
@@ -325,7 +328,7 @@
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <ProjectExtensions>
     <VisualStudio>
-      <UserProperties BuildVersion_BuildVersioningStyle="None.None.Increment.YearDayOfYear" BuildVersion_UseGlobalSettings="True" />
+      <UserProperties BuildVersion_UseGlobalSettings="True" BuildVersion_BuildVersioningStyle="None.None.Increment.YearDayOfYear" />
     </VisualStudio>
   </ProjectExtensions>
   <Import Project="$(SolutionDir)\.nuget\NuGet.targets" Condition="Exists('$(SolutionDir)\.nuget\NuGet.targets')" />
